pyproject.toml
[project]
name = "vrcphotoparser"
version = "0.1.0"
description = "parsing vrchat photo png chunk"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pillow>=12.1.1",
]
